+++ src/App.js
@@ -1,8 +1,8 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import LandingPage from "./components/LandingPage/LandingPage";
 import ResultsPage from "./components/ResultsPage/ResultsPage";
+import Login from "./components/UserAccountForm/UserAccountForm";
 import "./App.css";
-import SignUp from "./components/SignUp/SignUp";
 
 function App() {
   return (
@@ -10,8 +10,8 @@ function App() {
       <Routes>
         <Route index element={<LandingPage />} />
         <Route path="/results/:slug/:page/:chips/*" element={<ResultsPage />} />
-        <Route path="signup" element={<SignUp formStatus="signUp" />} />
-        <Route path="login" element={<SignUp formStatus="logIn" />} />
+        <Route path="signup" element={<Login formStatus="signUp" />} />
+        <Route path="login" element={<Login formStatus="logIn" />} />
       </Routes>
     </BrowserRouter>
   );
