--- src/components/Login/Login.module.css
@@ -1,71 +1,0 @@
-
-
-
-    .login {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        min-height: 200px;
-    }
-
-    .header {
-        margin-bottom: 15px;
-    }
-
-    .header > h4 {
-        font-size: 22px;
-    }
-
-    .header > p, .submit > p, .input > div > a, .submit > p > a {
-        font-size: 12px;
-    }
-
-    .submit > p > a, .header > h4 {
-        color: #4D8FBE;
-    }
-
-    .submit > p > a {
-        text-decoration: none;
-    }
-
-    .input {
-        display: flex;
-        flex-direction: column;
-        margin-bottom: 15px;
-    }
-
-    .input > input {
-        background-color: #F8F8F8;
-        border: 1px solid rgb(188, 188, 188);
-        height: 30px;
-        padding: 10px;
-    }
-
-    .input > div {
-        display: flex;
-        justify-content: flex-end;
-    }
-
-    .input > div > a {
-        color: rgb(168, 168, 168);
-        text-decoration: none;
-    }
-
-    .submit {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-top: 30px;
-    }
-    
-    .submit > input[type=submit] {
-        background-color: #4D8FBE;
-        border: none;
-        color: white;
-        width: 100px;
-        height: 32px;
-        border-radius: 40px;
-        margin-right: 10px;
-    }
