--- src/components/SignUp/SignUp.module.css
@@ -1,71 +1,0 @@
-.signup {
-height: 100vh;
-display: flex;
-flex-direction: column;
-align-items: center;
-background-color: #F4F2F0;
-}
-
-.signup > p {
-    text-align: center;
-    width: 100%;
-    padding: 24px 16px;
-    margin: 80px 80px 10px 80px;
-    font-size: 32px
-}
-
-.signup > div  {
-    border-radius: 10px;
-    padding: 30px;
-    background-color: white;
-}
-
-.signup > div > div > form, .signup > div > div {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    min-height: 80px;
-}
-
-.signup > div > div:last-of-type {
-    justify-content: center;
-}
-
-.signup > div > div > form > div {
-    margin-bottom: 30px;
-}
-
-.signup > div > div > form > div > p, .signup > div > div > form > div > p > a {
-    text-align: center;
-    font-size: 12px;
-    margin-top: 20px;
-}
-
-.signup > div > div > form > div > label {
-    display: block;
-    padding: 0px 0px 10px 0px;
-}
-
-.signup > div > div > form > div > input {
-    width: 360px;
-    height: 30px;
-}
-
-.signup > div > div > form > button {
-    border: none;
-    background-color: #4D8FBE;
-    color: white;
-    padding: 10px;
-    border-radius: 20px;
-    width: 360px;
-}
-
-
-.signup > div {
-width: 400px;
-}
-
-.blue {
-    color: #4D8FBE;
-    text-decoration: none;
-}
