+++ src/components/reusableComponents/ContainerPage/ContainerPage.module.css
@@ -1,0 +1,23 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: #F4F2F0;
+}
+
+.container > p {
+    text-align: center;
+    width: 100%;
+    padding: 24px 16px;
+    margin: 80px 80px 10px 80px;
+    font-size: 32px;
+}
+
+.container > div  {
+    margin-bottom: 250px;
+}
+
+.blue {
+    color: #4D8FBE;
+    text-decoration: none;
+}
