+++ src/components/reusableComponents/Form/Form.js
@@ -1,0 +1,8 @@
+import React from "react";
+import styles from "./Form.module.css";
+
+const Form = (props) => {
+  return <form className={styles.form}>{props.children}</form>;
+};
+
+export default Form;
