+++ src/components/reusableComponents/Form/Form.module.css
@@ -1,0 +1,47 @@
+.form {
+    padding: 30px;
+    width: 400px;
+    background-color: white;
+    border-radius: 10px;
+} 
+
+.form > button {
+    border: none;
+    background-color: #4D8FBE;
+    color: white;
+    padding: 10px;
+    border-radius: 20px;
+    width: 100%;
+}
+
+.form > div:last-of-type {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    min-height: 80px;
+    justify-content: center;
+}
+
+.form > header {
+    margin-bottom: 15px;
+}
+
+.form > header > h4 {
+    font-size: 22px;
+}
+
+.form > header > p, .submit > p{
+    font-size: 12px;
+}
+
+.form > header > h4 {
+    color: #4D8FBE;
+}
+
+:any-link {
+    text-decoration: none;
+}
+
+:any-link:hover {
+    text-decoration: underline;
+}
