+++ src/components/reusableComponents/InputWrapper/InputWrapper.js
@@ -1,0 +1,8 @@
+import React from "react";
+import styles from "./InputWrapper.module.css";
+
+const InputWrapper = (props) => {
+  return <div className={styles["input-wrapper"]}>{props.children}</div>;
+};
+
+export default InputWrapper;
